fixed original

+++ src/original-binomial.h
@@ -1,6 +1,7 @@
 #include <cstdint>
 
 namespace Binomial {
-double qlEuropeanCall(uint32_t steps, uint16_t expirationTime, double S, double K, double riskFreeRate, double volatility, double dividend_yield);
-double zubairEuropeanCall(uint32_t N, uint16_t expirationTime, double S, double K, double riskFreeRate, double volatility, double dividendYield);
+double qlEuropeanCall(int steps, int expirationTime, double S, double K, double riskFreeRate, double volatility, double dividend_yield);
+// double qlEuropeanCall(uint16_t steps, uint16_t expirationTime, double S, double K, double riskFreeRate, double volatility, double dividend_yield);
+double zubairEuropeanCall(int N, int expirationTime, double S, double K, double riskFreeRate, double volatility, double dividendYield);
 }  // namespace Binomial
